feat(vim): revert neovim lsp back to ale
The main issue i had with `neovim/nvim-lspconfig` that i could
not make it properly work to have eslint messages in the diagnostics.
It was either typescript or eslint but having both was not working all
the time.

Also i like more that whey the visual text in ale is displays. It shows
only when the cursor is on the line. I could not recreate this with
neovim's native lsp. Only has a popover but not as a virtual-text.